Handle case when FFmpeg encode isn't running and cancel is called.

+++ lib/active_encode/engine_adapters/ffmpeg_adapter.rb
@@ -98,10 +98,17 @@ module ActiveEncode
 
       # Cancel ongoing encode using pid file
       def cancel(id)
-        pid = get_pid(id)
-        Process.kill 'SIGTERM', pid.to_i
-
-        find id
+        encode = find id
+        if encode.running?
+          pid = get_pid(id)
+          Process.kill 'SIGTERM', pid.to_i
+          encode = find id
+        end
+        encode
+      rescue Errno::ESRCH
+        raise NotRunningError
+      rescue StandardError
+        raise CancelError
       end
 
     private
